Players
Changes made to character to allow for multiple alterations to skills
and inventory to reduce resizing of arrays. However, lits have to be
imported if changes are being made to the inventory or skills of a
player.
The readme has had conventions added to it for the programming of the
game.
Initial values for calculating the skills of the player have been added.